consumer/requirements.txt
boto3
kafka-python
python-dotenv
pandas
